backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pgvector==0.2.4
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# NLP and ML
transformers==4.35.2
torch==2.1.1
sentence-transformers==2.2.2
jieba==0.42.1
langdetect==1.0.9
spacy==3.7.2
opencc-python-reimplemented==0.1.7

# Data processing
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2

# HTTP and async
httpx==0.25.1
aiohttp==3.9.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
faker==20.1.0

# Email and notifications
emails==0.6
jinja2==3.1.2

# Cron and scheduling
apscheduler==3.10.4

# Finance data
yfinance==0.2.32
